requirements.txt
streamlit==1.38.0
pandas==2.2.2
openai==1.35.0
langchain-openai==0.1.8
langchain-community==0.2.6
python-dotenv==1.0.1